+++ module-apps/application-messages/windows/OptionsMessages.cpp
@@ -14,62 +14,55 @@
 
 using namespace style::window;
 
-std::list<gui::Item *> smsWindowOptions(app::ApplicationMessages *app, const SMSRecord &record)
+std::list<gui::Option> smsWindowOptions(app::ApplicationMessages *app, const SMSRecord &record)
 {
     ContactRecord contact = DBServiceAPI::ContactGetByIDWithTemporary(app, record.contactID)->front();
-    std::list<gui::Item *> options;
+    std::list<gui::Option> options;
 
     if (record.type == SMSType::FAILED) {
-        options.push_back(gui::Option{
-            utils::localize.get("sms_resend_failed"), [=, &record](gui::Item &item) {
-                app->resendSms(record);
-                app->returnToPreviousWindow();
-                return true;
-            }}.build());
-    }
-    options.push_back(gui::options::call(app, app::CallOperation::ExecuteCall, contact).build());
-    auto contactOperation = contact.isTemporary() ? app::ContactOperation::Add : app::ContactOperation::Details;
-    options.push_back(gui::options::contact(app, contactOperation, contact).build());
-    options.push_back(gui::Option{
-        UTF8(utils::localize.get("sms_forward_message")), [=](gui::Item &item) {
-            std::unique_ptr<gui::SwitchData> data = std::make_unique<SMSTextData>(record.body);
-            app->switchWindow(gui::name::window::new_sms, std::move(data));
-            return true;
-        }}.build());
-    options.push_back(gui::Option{
-        UTF8(utils::localize.get("sms_copy")), [=](gui::Item &item) {
-            Clipboard::getInstance().copy(record.body);
+        options.emplace_back(utils::localize.get("sms_resend_failed"), [=, &record](gui::Item &item) {
+            app->resendSms(record);
             app->returnToPreviousWindow();
             return true;
-        }}.build());
-    options.push_back(gui::Option{
-        UTF8(utils::localize.get("sms_delete_message")), [=](gui::Item &item) {
-            return app->removeSms(record);
-        }}.build());
+        });
+    }
+    options.push_back(gui::options::call(app, app::CallOperation::ExecuteCall, contact));
+    auto contactOperation = contact.isTemporary() ? app::ContactOperation::Add : app::ContactOperation::Details;
+    options.push_back(gui::options::contact(app, contactOperation, contact));
+    options.emplace_back(UTF8(utils::localize.get("sms_forward_message")), [=](gui::Item &item) {
+        std::unique_ptr<gui::SwitchData> data = std::make_unique<SMSTextData>(record.body);
+        app->switchWindow(gui::name::window::new_sms, std::move(data));
+        return true;
+    });
+    options.emplace_back(UTF8(utils::localize.get("sms_copy")), [=](gui::Item &item) {
+        Clipboard::getInstance().copy(record.body);
+        app->returnToPreviousWindow();
+        return true;
+    });
+    options.emplace_back(UTF8(utils::localize.get("sms_delete_message")),
+                         [=](gui::Item &item) { return app->removeSms(record); });
 
     return options;
 }
 
-std::list<gui::Item *> newMessageWindowOptions(app::ApplicationMessages *app,
+std::list<gui::Option> newMessageWindowOptions(app::ApplicationMessages *app,
                                                const std::string &requestingWindow,
                                                gui::Text *text)
 {
-    std::list<gui::Item *> options;
+    std::list<gui::Option> options;
 
-    options.push_back(gui::Option{
-        UTF8(utils::localize.get("sms_use_template")), [=](gui::Item &item) {
-            std::unique_ptr<gui::SwitchData> data = std::make_unique<SMSTemplateRequest>(requestingWindow);
-            app->switchWindow(gui::name::window::sms_templates, std::move(data));
-            return true;
-        }}.build());
+    options.emplace_back(UTF8(utils::localize.get("sms_use_template")), [=](gui::Item &item) {
+        std::unique_ptr<gui::SwitchData> data = std::make_unique<SMSTemplateRequest>(requestingWindow);
+        app->switchWindow(gui::name::window::sms_templates, std::move(data));
+        return true;
+    });
 
     if (Clipboard::getInstance().gotData()) {
-        options.push_back(gui::Option{
-            utils::localize.get("sms_paste"), [=](gui::Item &item) {
-                text->addText(Clipboard::getInstance().paste());
-                app->returnToPreviousWindow();
-                return true;
-            }}.build());
+        options.emplace_back(utils::localize.get("sms_paste"), [=](gui::Item &item) {
+            text->addText(Clipboard::getInstance().paste());
+            app->returnToPreviousWindow();
+            return true;
+        });
     }
 
     return options;